requirements.txt
torch==1.13.1+cpu
pyannote.audio==1.1.1
tqdm
ffmpeg-python
git+https://github.com/openai/whisper.git
python-dotenv